move EventItem to it's own file
